Fix old migration

+++ db/migrate/20120403174513_drop_old_tables.rb
@@ -1,9 +1,10 @@
 class DropOldTables < ActiveRecord::Migration
   def up
-    drop_table :interests
-    drop_table :mo_messages
-    drop_table :mt_messages
-    drop_table :user_interests
+    connection = ActiveRecord::Base.connection
+    drop_table :interests if connection.table_exists?(:interests)
+    drop_table :mo_messages if connection.table_exists?(:mo_messages)
+    drop_table :mt_messages if connection.table_exists?(:mt_messages)
+    drop_table :user_interests if connection.table_exists?(:user_interests)
   end
 
   def down
